adjusted database schema into postgresql #2

+++ src/main/resources/db/migration/postgresql/V1_1_0__init_database.sql
@@ -59,7 +59,7 @@ CREATE TABLE PRODUCTS_IMAGE_ASSETS (
 );
 
 CREATE TABLE APP_USER (
-                          ID SERIAL PRIMARY KEY,
+                          ID BIGINT PRIMARY KEY,
                           FIRST_NAME VARCHAR(250) NOT NULL,
                           LAST_NAME VARCHAR(250) NOT NULL,
                           USER_NAME VARCHAR(100) NOT NULL UNIQUE,
